# openapi.yaml
openapi: 3.0.1
info:
    title: Pratica 4 API
    description: Pratica 4 API documentation
    termsOfService: ''
    contact:
        name: ''
        url: ''
        email: 'pedpao@pedpao.com'
    license:
        name: unlicensed
        url: ''
    version: 0.0.1
servers:
    - url: http://localhost:8080
      description: Generated server url
paths:
...
    /api/informacoes-adicionais/{id}:
        get:
            tags:
                - informacoes-adicionais-resource
            operationId: getInformacoesAdicionais
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/InformacoesAdicionaisDTO'
        put:
            tags:
                - informacoes-adicionais-resource
            operationId: updateInformacoesAdicionais
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/InformacoesAdicionaisDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/InformacoesAdicionaisDTO'
        delete:
            tags:
                - informacoes-adicionais-resource
            operationId: deleteInformacoesAdicionais
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
        patch:
            tags:
                - informacoes-adicionais-resource
            operationId: partialUpdateInformacoesAdicionais
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/InformacoesAdicionaisDTO'
                    application/merge-patch+json:
                        schema:
                            $ref: '#/components/schemas/InformacoesAdicionaisDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/InformacoesAdicionaisDTO'
...
    /api/informacoes-adicionais:
        get:
            tags:
                - informacoes-adicionais-resource
            operationId: getAllInformacoesAdicionais
            parameters:
                - name: page
                  in: query
                  description: Zero-based page index (0..N)
                  required: false
                  schema:
                      minimum: 0
                      type: integer
                      default: 0
                - name: size
                  in: query
                  description: The size of the page to be returned
                  required: false
                  schema:
                      minimum: 1
                      type: integer
                      default: 20
                - name: sort
                  in: query
                  description: >-
                      Sorting criteria in the format: property,(asc|desc). Default sort
                      order is ascending. Multiple sort criteria are supported.
                  required: false
                  schema:
                      type: array
                      items:
                          type: string
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/InformacoesAdicionaisDTO'
        post:
            tags:
                - informacoes-adicionais-resource
            operationId: createInformacoesAdicionais
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/InformacoesAdicionaisDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/InformacoesAdicionaisDTO'
...
components:
    schemas:
...
        UserDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                login:
                    type: string
...
        InformacoesAdicionaisDTO:
            required:
                - bairro
                - cep
                - cpf
                - estado
                - id
                - numero
                - rua
                - telefone
            type: object
            properties:
                id:
                    minimum: 1
                    type: integer
                    format: int64
                telefone:
                    maxLength: 15
                    minLength: 10
                    type: string
                cpf:
                    maxLength: 11
                    minLength: 11
                    type: string
                cep:
                    maxLength: 8
                    minLength: 8
                    type: string
                rua:
                    maxLength: 255
                    minLength: 3
                    type: string
                bairro:
                    maxLength: 255
                    minLength: 3
                    type: string
                estado:
                    maxLength: 2
                    minLength: 2
                    type: string
                numero:
                    maxLength: 255
                    minLength: 1
                    type: string
                complemento:
                    type: string
                user:
                    $ref: '#/components/schemas/UserDTO'
...